refactor: merge directrendertasks
