+++ src/provider/index.ts
@@ -3,4 +3,4 @@ export { default as LocalDataProvider } from './LocalDataProvider';
 export * from './LocalDataProvider';
 export { deriveColors, deriveColumnDescriptions, exportRanking, exportTable } from './utils';
 export * from './interfaces';
-export { tasksAll } from './tasks';
+export { tasksAll } from './DirectRenderTasks';
